# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FoxRunner API",
    "description": "API de pilotage du scheduler et des scenarios FoxRunner.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/users/me/features": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "My Features",
        "operationId": "my_features_api_v1_users_me_features_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeatureFlagsPayload"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FeatureFlagsPayload": {
        "properties": {
          "features": {
            "additionalProperties": {
              "type": "boolean"
            },
            "type": "object",
            "title": "Features"
          }
        },
        "type": "object",
        "required": [
          "features"
        ],
        "title": "FeatureFlagsPayload"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/jwt/login"
          }
        }
      }
    }
  },
...
}